refactor numpy tests
