+++ python/tests/e2e/types/utils.py
@@ -1,6 +1,12 @@
 """Utility functions for type tests."""
 
 from collections.abc import Iterable
+from math import isinf, isnan
+
+
+# Minimum normalized positive value (smallest normal number)
+# Used for tolerance selection in float comparisons
+FLOAT_MIN_NORMAL = 2.2250738585072014e-308
 
 
 def assert_types(values: Iterable, expected_type: type, can_be_none: bool = False) -> None:
@@ -17,3 +23,52 @@ def assert_types(values: Iterable, expected_type: type, can_be_none: bool = Fals
         assert isinstance(value, expected_type), (
             f"Value at index {i} should be {expected_type.__name__}, got {type(value).__name__}"
         )
+
+
+def assert_float_equal(actual: float, expected: float | None, msg: str = "") -> None:
+    """Assert two float values are equal within appropriate tolerance.
+
+    Selects comparison strategy based on IEEE 754 value magnitude:
+    - Subnormal (|x| < 2.2e-308): absolute tolerance 1e-325
+    - Large (|x| > 1e10): relative tolerance 1e-14
+    - Regular: absolute tolerance 1e-10
+    """
+    error_msg = msg or f"Expected {expected}, got {actual}"
+
+    # None
+    if expected is None:
+        assert actual is None, error_msg
+        return
+    # NaN
+    if isnan(expected):
+        assert isnan(actual), error_msg
+        return
+    # inf, -inf
+    if isinf(expected):
+        assert actual == expected, error_msg
+        return
+
+    abs_expected = abs(expected)
+    diff = abs(actual - expected)
+
+    # Subnormal range (very small numbers near minimum representable)
+    if abs_expected < FLOAT_MIN_NORMAL:
+        assert diff <= 1e-325, error_msg
+    # Large values - use relative tolerance for ~15 digit precision
+    elif abs_expected > 1e10:
+        assert diff <= abs_expected * 1e-14, error_msg
+    # Regular values - absolute tolerance
+    else:
+        assert diff < 1e-10, error_msg
+
+
+def assert_floats_equal(actual: Iterable[float], expected: Iterable[float]) -> None:
+    """Assert two iterables of float values are equal element-wise.
+
+    Calls assert_float_equal on consecutive elements.
+    """
+    actual_list = list(actual)
+    expected_list = list(expected)
+    assert len(actual_list) == len(expected_list), f"Length mismatch: {len(actual_list)} != {len(expected_list)}"
+    for i, (a, e) in enumerate(zip(actual_list, expected_list)):
+        assert_float_equal(a, e, f"Mismatch at index {i}: expected {e}, got {a}")
